backend/tests/test_auth_middleware.py
"""
Tests for Authentication Middleware
"""
import pytest
from fastapi import HTTPException
from app.core.auth_middleware import get_current_user, get_current_user_optional
from fastapi.security import HTTPAuthorizationCredentials
import jwt
from datetime import datetime, timedelta, timezone
from app.core.config import settings


class TestAuthMiddleware:
    """Test authentication middleware functionality"""
    
    def test_valid_token_extraction(self):
        """Test extracting user_id from valid JWT token"""
        # Create valid token
        user_id = "test-user-123"
        token_data = {
            "user_id": user_id,
            "exp": datetime.now(timezone.utc) + timedelta(hours=1)
        }
        token = jwt.encode(token_data, settings.SECRET_KEY, algorithm="HS256")
        
        # This would require async test setup with FastAPI dependency injection
        # For now, we verify token creation works
        assert token is not None
        assert isinstance(token, str)
    
    def test_expired_token_detection(self):
        """Test that expired tokens are rejected"""
        user_id = "test-user-123"
        token_data = {
            "user_id": user_id,
            "exp": datetime.now(timezone.utc) - timedelta(hours=1)  # Expired
        }
        token = jwt.encode(token_data, settings.SECRET_KEY, algorithm="HS256")
        
        # Verify token is expired
        with pytest.raises(jwt.ExpiredSignatureError):
            jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
    
    def test_invalid_token_signature(self):
        """Test that tokens with wrong signature are rejected"""
        user_id = "test-user-123"
        token_data = {
            "user_id": user_id,
            "exp": datetime.now(timezone.utc) + timedelta(hours=1)
        }
        # Sign with wrong key
        token = jwt.encode(token_data, "wrong-secret-key", algorithm="HS256")
        
        # Verify signature is invalid
        with pytest.raises(jwt.InvalidTokenError):
            jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
    
    def test_missing_user_id_in_token(self):
        """Test that tokens without user_id are rejected"""
        token_data = {
            "some_other_field": "value",
            "exp": datetime.now(timezone.utc) + timedelta(hours=1)
        }
        token = jwt.encode(token_data, settings.SECRET_KEY, algorithm="HS256")
        
        # Decode and verify user_id is missing
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
        assert "user_id" not in payload
    
    def test_token_creation_with_correct_algorithm(self):
        """Test JWT creation uses HS256 algorithm"""
        user_id = "test-user-123"
        token_data = {
            "user_id": user_id,
            "exp": datetime.now(timezone.utc) + timedelta(hours=1)
        }
        token = jwt.encode(token_data, settings.SECRET_KEY, algorithm="HS256")
        
        # Decode and verify
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
        assert payload["user_id"] == user_id


@pytest.mark.asyncio
class TestAuthMiddlewareAsync:
    """Async tests for auth middleware (requires FastAPI test client)"""
    
    async def test_placeholder_async(self):
        """Placeholder for future async tests"""
        # Future: Add FastAPI TestClient integration tests
        assert True
