backend/requirements-windows.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==24.3.0
Authlib==1.6.5
backoff==2.2.1
bandit==1.8.6
bcrypt==5.0.0
black==25.9.0
boto3==1.35.82
botocore==1.35.82
build==1.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.3.0
coloredlogs==15.0.1
cryptography==46.0.2
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dparse==0.6.4
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.6
fastuuid==0.13.5
filelock==3.19.1
flake8==7.3.0
flatbuffers==25.9.23
frozenlist==1.5.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.183.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.2.4
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.5.2
jq==1.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==34.1.0
limits==5.6.0
litellm==1.77.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
motor==3.3.2
mpmath==1.3.0
multidict==6.1.0
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.2.1
oauthlib==3.2.2
onnxruntime==1.23.0
openai==1.99.9
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
posthog==5.4.0
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.6.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.0
pypdf==6.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.13.2
s3transfer==0.10.4
s5cmd==0.2.0
safetensors==0.6.2
safety==3.6.2
safety-schemas==0.0.16
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.41.3
stevedore==5.5.0
stripe==12.3.0
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.37.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.3
yarl==1.18.3
zipp==3.22.0
PyGithub==2.1.1
PyNaCl==1.6.0
prometheus-client==0.20.0
psycopg2-binary==2.9.10
redis==5.2.1
# Note: Mono C# compiler (mcs) installed via apt-get, not pip
cryptography==46.0.2
sse-starlette==2.1.3
reportlab==4.4.4
weasyprint==66.0
